requirement.txt
pytesseract
pyqt5
pillow
pyinstaller